Revert "treltexx: Configure MDNIE to auto by default"
While this does indeed improve color accuracy, it does so at the
cost of heatings and general performance issues.
I wasn't able to narrow this down using standard Linux analysis tools
(top and friends), but we know MDNIE changes (as in rearranges) the
liquid crystals in the AMOLED display, so this probably causing issues
in MDNIE 'auto' mode because it's constantly adjusting itself.

This reverts commit bfa987b4e0f4bddef52b4cb7d4d3d5f6c60d66e0.

Change-Id: I901bd2e41ba32264a088c3b8efe29ece53ac41ed