treltexx: Enforce incall wideband audio
There seems to be a bug in the current implementation causing earpiece
output to be muted in the first call. After switching the output to
speakerphone and back, earpiece audio output is fine.

My current guess is that the RIL WB_AMR callback is not working correctly
right now, so that needs to be examined in the future.

Change-Id: If6d32b48a70ea1904b99ccded2e9c2bd961a6355